requirements.txt
django-ckeditor==5.4.0
django-countries==5.0
Pillow==4.3.0
imagekit==1.0.6
django-cleanup
